refactor(ui): make confirmation dialog reusable HBL-250 #time 10m
