+++ libs/shared/models/dto/src/lib/EventAppointment/index.ts
@@ -1,3 +1,3 @@
-import EventAppointmentDTO from './EventAppointment';
+import EventDTO from './EventAppointment';
 
-export default EventAppointmentDTO;
+export default EventDTO;
