+++ libs/ui/components/src/lib/Dialogs/ConfirmationDialog/Appointment/index.ts
@@ -1,3 +1,3 @@
 import Appointment from './Appointment';
 
-export { Appointment };
+export default Appointment;
