--- libs/ui/components/src/lib/Dialogs/ConfirmationDialog/CalendarAppointment/index.ts
@@ -1,3 +1,0 @@
-import CalendarAppointment from './CalendarAppointment';
-
-export default CalendarAppointment;
